+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_polling.py
@@ -17,6 +17,12 @@ if TYPE_CHECKING:
     from azure.core.pipeline import PipelineResponse
 
 
+def raise_error(response, errors, message):
+    for err in errors:
+        message += "({}) {}\n".format(err["code"], err["message"])
+    raise HttpResponseError(message=message, response=response)
+
+
 class TrainingPolling(LocationPolling):
     """Polling method overrides for training endpoints.
     """
@@ -40,6 +46,14 @@ class TrainingPolling(LocationPolling):
             status = body['modelInfo']['status']
             if not status:
                 raise BadResponse("No status found in body")
+            if status.lower() == "invalid":
+                train_result = body.get('trainResult')
+                if train_result:
+                    errors = train_result.get("errors")
+                    if errors:
+                        message = "Invalid model created with ID={}\n".format(body["modelInfo"]["modelId"])
+                        raise_error(response, errors, message)
+                return "Failed"
             if status.lower() != "creating":
                 return "Succeeded"
 
@@ -50,8 +64,6 @@ class TrainingPolling(LocationPolling):
 
 class AnalyzePolling(OperationResourcePolling):
     """Polling method overrides for custom analyze endpoints.
-
-    :param str operation_location_header: Name of the header to return operation format (default 'operation-location')
     """
 
     def get_status(self, pipeline_response):  # pylint: disable=no-self-use
@@ -78,10 +90,7 @@ class AnalyzePolling(OperationResourcePolling):
             if analyze_result:
                 errors = analyze_result.get("errors")
                 if errors:
-                    message = ""
-                    for err in errors:
-                        message += "({}) {}\n".format(err.get("code"), err.get("message"))
-                    raise HttpResponseError(message)
+                    raise_error(response, errors, message="")
         return status